Add sort functionality
Can now sort by name, cost, power, pool, or owned. Also reformatted the
header row of the table to be generated by a for loop using new context
added to the view. There are some commented out lines in urls, views,
and index that should be removed later if not used.